fix code smells
